test(slack): colocate oauth bind wiring tests
